requirements.txt
## CORE REQUIREMENTS
Django==2.0.4                       # Web framework
psycopg2-binary==2.7.4              # Postgres connector
dj-database-url==0.5.0              # DB config via ENV vars
pytz==2018.3                        # Timezone support
django-tz-detect==0.2.9             # Timezone support on templates
django-extensions==2.0.6            # Various Django extras
ipython==6.3.1                      # Better Django shell experience
django-allauth==0.35.0              # Auth/Registration
django-formtools==2.1               # Wizard-type forms
django-dotenv==1.4.2                # To read .env file (because this project is open source)

## INFRASTRUCTURE REQUIREMENTS
boto3==1.7.2                        # to talk with amazon
django-storages==1.6.6              # to upload to S3
Pillow==5.1.0                       # to deal with images

## FRONTEND REQUIREMENTS
django-compressor==2.2              # asset compressor
django-bootstrap4==0.0.6            # bootstrap4 forms
django-related-select==0.9.1        # for related select fields on forms with AJAX
django-tinymce==2.7.0               # for HTML editor for examples

## ASYNC TASKS
celery==4.1.0                       # Async Task Framework

## PRODUCTION REQUIREMENTS
whitenoise==3.3.0                   # Serve static files via app server
gevent==1.2.2                       # Event-driven web server patch
gunicorn==19.7.1                    # Threaded WSGI server
hiredis==0.2.0                      # C-based redis driver
django-redis-cache==1.7.1           # Redis connector
raven==6.6.0                        # Sentry connector
newrelic==3.2.0.91                  # New Relic connector
